+++ lib/markerDetails/marker_details.dart
@@ -1,4 +1,4 @@
-import 'package:cracker_app/+widgets/marker_map.dart';
+import 'package:cracker_app/markerDetails/widgets/marker_map.dart';
 import 'package:flutter/material.dart';
 import 'package:geolocator/geolocator.dart';
 
@@ -10,16 +10,14 @@ class MarkerDetails extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final imageFilename = marker['imageFilename'];
     return Scaffold(
       appBar: AppBar(),
       body: Column(
         children: [
-          FlatButton(
-            child: Text('Go back!'),
-            onPressed: () {
-              Navigator.pop(context);
-            },
-          ),
+          imageFilename != null
+              ? Image.network('https://cracker.red/images/$imageFilename')
+              : Text(''),
           Text(marker['polish']['name']),
           Text(marker['polish']['description']),
           Text(marker['english']['name']),
